+++ ocf_datapipes/convert/__init__.py
@@ -1,12 +1,18 @@
 """Various conversion datapipes"""
-from ocf_datapipes.convert.numpy.batch.gsp import (
+from ocf_datapipes.convert.numpy_batch.gsp import (
     ConvertGSPToNumpyBatchIterDataPipe as ConvertGSPToNumpyBatch,
 )
-from ocf_datapipes.convert.numpy.batch.pv import (
+from ocf_datapipes.convert.numpy_batch.pv import (
     ConvertPVToNumpyBatchIterDataPipe as ConvertPVToNumpyBatch,
 )
-from ocf_datapipes.convert.numpy.batch.sensor import (
+from ocf_datapipes.convert.numpy_batch.sensor import (
     ConvertSensorToNumpyBatchIterDataPipe as ConvertSensorToNumpyBatch,
+)
+from ocf_datapipes.convert.numpy_batch.nwp import (
+    ConvertNWPToNumpyBatchIterDataPipe as ConvertNWPToNumpyBatch,
+)
+from ocf_datapipes.convert.numpy_batch.satellite import (
+    ConvertSatelliteToNumpyBatchIterDataPipe as ConvertSatelliteToNumpyBatch,
 )
 
 from .coordinates import ConvertGeostationaryToLonLatIterDataPipe as ConvertGeostationaryToLonLat
@@ -14,5 +20,3 @@ from .coordinates import ConvertLonLatToOSGBIterDataPipe as ConvertLonLatToOSGB
 from .coordinates import ConvertOSGBToLonLatIterDataPipe as ConvertOSGBToLonLat
 from .numpy.gsp import ConvertGSPToNumpyIterDataPipe as ConvertGSPToNumpy
 from .numpy.pv import ConvertPVToNumpyIterDataPipe as ConvertPVToNumpy
-from .nwp import ConvertNWPToNumpyBatchIterDataPipe as ConvertNWPToNumpyBatch
-from .satellite import ConvertSatelliteToNumpyBatchIterDataPipe as ConvertSatelliteToNumpyBatch
