+++ ocf_datapipes/convert/numpy_batch/satellite.py
